# test/openapi.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/api/auth/login": {
            "post": {
                "operationId": "Auth_login",
                "summary": "登录",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseDto"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "auth"
                ]
            },
            "get": {
                "operationId": "Auth_loginByAuth0",
                "summary": "基于 Auth0 登录",
                "parameters": [
                    {
                        "name": "redirect",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "auth"
                ]
            }
        },
        "/api/auth/register": {
            "get": {
                "operationId": "Auth_registerByAuth0",
                "summary": "基于 Auth0 注册",
                "parameters": [
                    {
                        "name": "redirect",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "auth"
                ]
            },
            "post": {
                "operationId": "Auth_register",
                "summary": "注册",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "auth"
                ]
            }
        },
        "/api/auth/callback": {
            "post": {
                "operationId": "Auth_callback",
                "summary": "处理 Auth0 回调",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Auth0CallbackData"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "auth"
                ]
            }
        },
...
    },
    "info": {
        "title": "RSS Impact server docs",
        "description": "RSS Impact server docs",
        "version": "0.1.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "securitySchemes": {
            "cookie": {
                "type": "apiKey",
                "in": "cookie",
                "name": "connect.sid"
            }
        },
        "schemas": {
            "ResponseDto": {
                "type": "object",
                "properties": {
                    "statusCode": {
                        "type": "number"
                    },
                    "message": {
                        "type": "string"
                    },
                    "error": {
                        "type": "string"
                    },
                    "stack": {
                        "type": "string"
                    },
                    "data": {
                        "type": "object"
                    }
                }
            },
...
            "User": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "number",
                        "title": "ID",
                        "example": 1
                    },
                    "createdAt": {
                        "format": "date-time",
                        "type": "string",
                        "title": "创建时间",
                        "example": "2023-12-31T16:00:00.000Z"
                    },
                    "updatedAt": {
                        "format": "date-time",
                        "type": "string",
                        "title": "更新时间",
                        "example": "2023-12-31T16:00:00.000Z"
                    },
                    "username": {
                        "type": "string",
                        "title": "用户名",
                        "example": "admin"
                    },
                    "password": {
                        "type": "string",
                        "title": "密码",
                        "example": "123456"
                    },
                    "email": {
                        "type": "string",
                        "title": "邮箱",
                        "example": "admin@example.com"
                    },
                    "emailVerified": {
                        "type": "boolean",
                        "title": "邮箱已验证",
                        "example": true
                    },
                    "disablePasswordLogin": {
                        "type": "boolean",
                        "title": "禁用密码登录",
                        "example": false
                    },
                    "avatar": {
                        "type": "string",
                        "title": "头像",
                        "example": "URL_ADDRESS"
                    },
                    "roles": {
                        "title": "角色",
                        "example": [
                            "admin"
                        ],
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "accessToken": {
                        "type": "string",
                        "title": "接口访问令牌",
                        "description": "接口访问令牌，部分情况可替代账号密码",
                        "example": "rss-impact:fef26d6e-040f-4a7b-8d6a-4e4f12e107b6"
                    },
                    "auth0Id": {
                        "type": "string",
                        "title": "Auth0 ID",
                        "description": "绑定的 auth0 账号",
                        "example": "github|114514"
                    }
                },
                "required": [
                    "id",
                    "createdAt",
                    "updatedAt",
                    "username",
                    "email",
                    "emailVerified",
                    "disablePasswordLogin",
                    "roles",
                    "accessToken"
                ]
            },
...
            "LoginDto": {
                "type": "object",
                "properties": {
                    "username": {
                        "type": "string",
                        "description": "用户名",
                        "example": "admin"
                    },
                    "password": {
                        "type": "string",
                        "minLength": 6,
                        "description": "密码",
                        "example": "123456"
                    }
                },
                "required": [
                    "username",
                    "password"
                ]
            },
            "Auth0CallbackData": {
                "type": "object",
                "properties": {
                    "id_token": {
                        "type": "string"
                    },
                    "state": {
                        "type": "string"
                    },
                    "sid": {
                        "type": "string"
                    }
                },
                "required": [
                    "id_token"
                ]
            },
            "RegisterDto": {
                "type": "object",
                "properties": {
                    "username": {
                        "type": "string",
                        "minLength": 0,
                        "maxLength": 128,
                        "description": "用户名",
                        "example": "user"
                    },
                    "password": {
                        "type": "string",
                        "minLength": 6,
                        "maxLength": 128,
                        "description": "密码",
                        "example": "123456"
                    },
                    "email": {
                        "type": "string",
                        "minLength": 0,
                        "maxLength": 128,
                        "description": "邮箱",
                        "example": "user@example.com"
                    }
                },
                "required": [
                    "username",
                    "password",
                    "email"
                ]
            },
...
        }
    }
}